requirements.txt
pytest==8.4.*
pytest-django==4.11.*
coverage==7.*
